lmao, best results so far 3k and trying to get more now
Every time the model before this executes a buy order, it puts half of the btc in a vault that it can't touch. Long story short, 3K up after like 6 months on the training set (portfolio value). Had like 45 bucks in cash so it just bought btc until it had no money, was forced to hold it, and that's been the best way to do things so far lol